+++ app-demo/build.gradle
@@ -8,7 +8,7 @@ android {
     compileSdk 32
 
     defaultConfig {
-        applicationId "com.xieql.logger"
+        applicationId "com.october.logger"
         minSdk 21
         targetSdk 32
         versionCode 1
