+++ app-demo/src/main/java/com/october/logger/App.kt
@@ -1,0 +1,16 @@
+package com.october.logger
+
+import android.app.Application
+
+
+class App: Application() {
+
+    override fun onCreate() {
+        super.onCreate()
+
+        TestLoggerHelper.init(this)
+    }
+
+}
+
+
