--- app-demo/src/main/java/com/xieql/logger/App.kt
@@ -1,16 +1,0 @@
-package com.xieql.logger
-
-import android.app.Application
-
-
-class App: Application() {
-
-    override fun onCreate() {
-        super.onCreate()
-
-        TestLoggerHelper.init(this)
-    }
-
-}
-
-
