+++ otLogger/src/main/java/com/october/lib/logger/util/Debug.kt
@@ -1,0 +1,5 @@
+package com.october.lib.logger.util
+
+internal fun debugLog(msg:String){
+    //System.out.println("Logger调试日志：${msg}")
+}
