--- otLogger/src/main/java/com/xieql/lib/logger/print/IPrinter.kt
@@ -1,7 +1,0 @@
-package com.xieql.lib.logger.print
-
-import com.xieql.lib.logger.LogLevel
-
-interface IPrinter {
-    open fun print(logLevel: LogLevel, tag:String?, msg:String?, thr: Throwable?, param:Any?)
-}
