--- otLogger/src/main/java/com/xieql/lib/logger/util/Debug.kt
@@ -1,5 +1,0 @@
-package com.xieql.lib.logger.util
-
-internal fun debugLog(msg:String){
-    System.out.println("Logger调试日志：${msg}")
-}
